+++ src/js/index.js
@@ -1,4 +1,8 @@
+import { addDefaultTopMenStyle, initToggleTopMenu } from './topMenu'
 import { addDefaultSidebarStyle, initToggleSidebar } from './sidebar'
+
+/* ちらつき防止のため、topMenuの初期スタイルを追加する */
+addDefaultTopMenStyle()
 
 /* ちらつき防止のため、sidebarの初期スタイルを追加する */
 addDefaultSidebarStyle()
@@ -31,135 +35,18 @@ function addFeedbackLink() {
   topMenuNav.appendChild(li)
 }
 
-
-/* topMenuの折りたたみ等 */
-const topMenuOpenStyle = `
-  #wrapper {
-    margin-left: 180px
-  }
-
-  #top-menu {
-    width: 180px;
-  }
-`
-
-const topMenuCloseStyle = `
-  #wrapper {
-    margin-left: 52px;
-  }
-
-  #top-menu {
-    width: 52px;
-  }
-`
-
-function addDefaultTopMenStyle() {
-  const styleTag = document.createElement('style')
-  const isTopMenuOpen = localStorage.getItem('isTopMenuOpen') === 'true'
-  styleTag.setAttribute('data-style', 'topMenu')
-  styleTag.textContent = isTopMenuOpen ? topMenuOpenStyle : topMenuCloseStyle
-
-  document.head.appendChild(styleTag)
-}
-addDefaultTopMenStyle()
-
-function updateTopMenuStyle() {
-  const styleTag = document.querySelector('style[data-style="topMenu"]')
-  const isTopMenuOpen = localStorage.getItem('isTopMenuOpen') === 'true'
-  styleTag.textContent = isTopMenuOpen ? topMenuOpenStyle : topMenuCloseStyle
-}
-
-function getTopMenu() {
-  return document.getElementById('top-menu')
-}
-
-function topMenuExists() {
-  return document.getElementById('top-menu') !== null
-}
-
-function isLoggedIn() {
-  return document.getElementById('loggedas') !== null
-}
-
-function addLogoutStyle() {
-  // ログアウト中は#loggedasが存在しないので、#accountにmargin-top: auto;を適用したい
-  if(!isLoggedIn()) {
-    const account = document.getElementById('account')
-    account.style.marginTop = 'auto'
-  }
-}
-
-function addBtnToToggleTopMenu() {
-  if(!topMenuExists()) return
-
-  const topMenu = getTopMenu()
-
-  // 枠を作成
-  const div = document.createElement('div')
-  div.style.cssText = `
-    border-bottom: 1px solid #ccd5d9;
-    margin-top: -53px;
-    margin-bottom: 1rem;
-    padding-bottom: 1rem;
-  `
-  div.classList.add('aw_toggleTopMenuWrap')
-
-  // ボタンを作成
-  const btn = document.createElement('button')
-  btn.textContent = 'メニューを固定する'
-  btn.style.cssText = `
-    display: block;
-    width: 100%;
-    background-repeat: no-repeat;
-    background-size: 20px 20px;
-    background-position: left center;
-    border: 0;
-  `
-  btn.classList.add('aw_toggleTopMenu')
-
-  toggleTopMenu(btn)
-
-  div.appendChild(btn)
-  topMenu.appendChild(div)
-}
-
-function toggleTopMenu(btn) {
-  let isTopMenuOpen = false
-  if(localStorage.getItem('isTopMenuOpen') === 'true') {
-    isTopMenuOpen = true
-  } else {
-    isTopMenuOpen = false
-  }
-
-  btn.addEventListener('click', () => {
-    isTopMenuOpen = !isTopMenuOpen
-
-    if(isTopMenuOpen) {
-      document.body.classList.add('isTopMenuOpen')
-      localStorage.setItem('isTopMenuOpen', true)
-    } else {
-      document.body.classList.remove('isTopMenuOpen')
-      localStorage.setItem('isTopMenuOpen', false)
-    }
-    updateTopMenuStyle()
-  })
-}
-
 /**
  * その他一般的な処理
  */
 window.addEventListener('DOMContentLoaded', () => {
-  addBtnToToggleTopMenu()
-  addLogoutStyle()
   removeLoggedasText()
   addFeedbackLink()
 
-  // topMenuの復原処理
-  if(localStorage.getItem('isTopMenuOpen') === 'true') {
-    document.body.classList.add('isTopMenuOpen')
-  } else {
-    document.body.classList.remove('isTopMenuOpen')
-  }
+  /**
+   * TopMenuの開閉機能
+   */
+  initToggleTopMenu()
+
 
   /**
    * Sidebarの開閉機能
